refactor!: rename modules
BREAKING CHANGE: breaks public imports
