+++ .github/workflows/ci.yml
@@ -38,7 +38,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        python: ['2.7', '3.3', '3.4', '3.5', '3.6', '3.7', '3.8', '3.9']
+        python: ['2.7', '3.3', '3.4', '3.5', '3.6', '3.7', '3.8', '3.9', '3.10']
         os: ['ubuntu-18.04', windows-latest]
         exclude:
           - os: windows-latest
