+++ pdb_attach/socket.pyi
