+++ test/test_detach.py
