+++ test/test_signal.py
